HOOKSに役割分断
